signup and signin and some bugs fix

+++ delete_post.php
@@ -1,8 +1,8 @@
 <?php
 include("database-config.php");
-
-$post_id = $_GET["id"];
-$con->query("DELETE FROM posts WHERE post_id='$post_id'");
+session_start();
+if (isset($_SESSION['user_id'])) {
+    $post_id = $_GET["id"];
+    $con->query("DELETE FROM posts WHERE post_id='$post_id'");
+}
 header("location:home.php");
-
-?>
